added useShowError and refactoring
